Move CLI implementation behind thin mech binary
