+++ Cargo.toml
@@ -22,6 +22,11 @@ name = "mechc"
 path = "src/bin/mechc.rs"
 required-features = ["compiler"]
 
+[[bin]]
+name = "mech"
+path = "src/bin/mech.rs"
+required-features = ["cli"]
+
 [profile.release]
 debug = false
 
@@ -41,10 +46,10 @@ base = ["baselib", "pretty_print", "serde", "compiler", "program", "mika",
       "formatter", "serve", "bundle_web", "run", "test", "repl", "build","whos", "async", "trace",
       "mech-core/base", "mech-program/base", "mech-syntax/base",
       ]
-build = ["compiler", "mech-runtime", "mech-runtime/default"]
-repl = []
+build = ["cli", "compiler", "mech-runtime", "mech-runtime/default"]
+repl = ["cli"]
 run = ["mech-program/program", "cli", "cli_host", "project"]
-test = ["variable_define", "invariant_define", "symbol_table", "bool", "mech-runtime", "mech-runtime/default"]
+test = ["cli", "variable_define", "invariant_define", "symbol_table", "bool", "mech-runtime", "mech-runtime/default"]
 cli = []
 project = ["mech-runtime", "mech-runtime/default"]
 cli_host = ["mech-runtime", "mech-runtime/default", "mech-host-cli", "mech-host-cli/provider"]
@@ -54,7 +59,7 @@ serve = ["cli", "project", "web_host", "formatter", "mech-runtime", "mech-runtim
 host_delegation = ["serde", "mech-runtime", "mech-runtime/host_delegation", "mech-host-browser", "mech-host-browser/delegation"]
 host_delegation_signing = ["host_delegation", "mech-runtime/host_delegation_signing", "mech-host-browser/delegation_signing"]
 whos = ["pretty_print", "variables", "symbol_table"]
-formatter = ["mech-syntax/formatter"]
+formatter = ["cli", "mech-syntax/formatter"]
 async = []
 trace = ["mech-program/trace"]
 linked_stdlib = ["mech-program/linked_stdlib"]
