+++ src/cli.rs
@@ -1,3 +1,13 @@
+#[cfg(any(
+  feature = "build",
+  feature = "formatter",
+  feature = "run",
+  feature = "serve",
+  feature = "bundle_web",
+  feature = "test",
+  feature = "repl"
+))]
+pub mod app;
 #[cfg(feature = "bundle_web")]
 pub mod bundle_web;
 #[cfg(any(feature = "serve", feature = "run"))]
